refactor(jazz-tools): split row_histories/apply.rs into resolution + mutations
Separate the pure visibility/merge math (frontier walk, latest common
ancestor, per-column merge, delete-winner, computed visible preview)
into resolution.rs from the storage-mutating verbs (apply_row_batch,
patch_row_batch_state) in mutations.rs. The previous apply.rs name
described the verb but not the module boundary, and roughly half its
contents were resolution helpers that never touched Storage.

Rename the internal RowBatchApply struct to AppliedRowBatch — it
captures before/after state, it isn't the operation itself.

Update mod.rs and types.rs imports; expanded module headers explain
the dependency direction (mutations → resolution; types also pulls
from resolution for VisibleRowEntry::rebuild_*).
